Introduce PSA_KEY_ID_T macro
Currently psa_key_id_t is equal to psa_key_extended_id_t
if MBEDTLS_PSA_CRYPTO_KEY_EXTENDED_ID configuration flag
is set.

The purpose of this commit and the following is for
psa_key_id_t to always be as defined by the PSA
Cryptography API.

To do so, this commit first replace any occurence in the code
of psa_key_id_t by the macro PSA_KEY_ID_T and defines it
to be equal to psa_key_id_t. The next commit will move
PSA_KEY_ID_T to be equal to psa_key_extended_id_t when the
MBEDTLS_PSA_CRYPTO_KEY_EXTENDED_ID configuration flag
is set.

By the way, this commit updates the key identifier initialization
in the tests to be compatible with a composit extended key
identifier.

Signed-off-by: Ronald Cron <ronald.cron@arm.com>