tests: psa: Reset systematically key attributes
After a call to psa_get_key_attributes() to retrieve
the attributes of a key into a psa_key_attributes_t
structure, a call to psa_reset_key_attributes() is
mandated to free the resources that may be
referenced by the psa_key_attributes_t structure.
Not calling psa_reset_key_attributes() may result in
a memory leak.

In PSA test cases, the call to psa_reset_key_attributes()
was missing in a fair amount of places which outlines
that it is quite easy to forget about it.

To try and avoid this to happen again in the future, this
commit reset systematically in the clean-up section of PSA
test cases the attributes structures used by the test cases
even. It does this even where it is not necessary to do it
(no call to psa_get_key_attributes() only calls to
psa_set_key_xxx APIs).

The hope with this systematic and ubiquitous reset is that
developers (of tests and beyond) will not forget about it.

Signed-off-by: Ronald Cron <ronald.cron@arm.com>